# docs/openapi.yml
openapi: "3.0.0"
info:
  title: The Fiagram API
  description: The REST Full API specification for the Fiagram project
  version: 1.0.0
servers:
  - url: http://api.fiagram.com/v1
    description: Production server

  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Sign up a new account
      operationId: signUp
      description: |
        Sign up account with user information
      security: [] # public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created
          headers:
            Set-Cookie:
              description: |
                Sets `refresh_token=<opaque>; HttpOnly; Secure; Path=/api/v1/auth/token; SameSite=Lax`.
              schema:
                type: string
                pattern: "^refresh_token="
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"

        "500": { $ref: "#/components/responses/InternalServerError" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /auth/signin:
    post:
      tags: [Auth]
      summary: Sign in
      description: |
        Validates credentials, sets refresh token to cookie, returns access token in body.
      operationId: signIn
      security: [] # public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigninRequest"
      responses:
        "200":
          description: Signed in successfully
          headers:
            Set-Cookie:
              description: |
                Sets `refresh_token=<opaque>; HttpOnly; Secure; Path=/api/v1/auth/token; SameSite=Lax`.
              schema:
                type: string
                pattern: "^refresh_token="
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SigninResponse"
        "500": { $ref: "#/components/responses/InternalServerError" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "404": { $ref: "#/components/responses/NotFound" }
  /auth/token/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using refresh token
      description: |
        Use a refresh token to obtain a new access token. Server MAY rotate refresh tokens.
        Refresh token can be provided via HttpOnly cookie.
      operationId: refreshToken
      security:
        - RefreshTokenCookie: []
      responses:
        "200":
          description: Tokens refreshed
          headers:
            Set-Cookie:
              description: |
                Sets `refresh_token=<opaque>; HttpOnly; Secure; Path=/api/v1/auth/token; SameSite=Lax`.
              schema:
                type: string
                pattern: "^refresh_token="
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /auth/token/signout:
    post:
      tags: [Auth]
      summary: Sign out and revoke refresh token
      description: >
        Revokes the `refresh_token`. If `refresh_token` is stored in cookie,
        this endpoint should also clear it.
      operationId: signOut
      security:
        - RefreshTokenCookie: []
      responses:
        "204":
          description: Signed out successfully
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # -------------------------------- Profile
  /profile/me:
    get:
      tags: [Profile]
      summary: Get current account information
      operationId: getProfileMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current account information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileMeResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

    put:
      tags: [Profile]
      summary: Update current account information
      operationId: updateProfileMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileMeRequest"
      responses:
        "200":
          description: Account information updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileMeResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profile/me/password:
    put:
      tags: [Profile]
      summary: Update current account password
      operationId: updateProfilePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordRequest"
      responses:
        "204":
          description: Password updated successfully
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profile/webhooks:
    get:
      tags: [Profile]
      summary: List the account's webhooks
      operationId: getProfileWebhooks
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: List of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

    post:
      tags: [Profile]
      summary: Create a new webhook
      operationId: createProfileWebhook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Webhook"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /profile/webhooks/{webhookId}:
    get:
      tags: [Profile]
      summary: Get a specific webhook
      operationId: getProfileWebhook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      responses:
        "200":
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

    put:
      tags: [Profile]
      summary: Update a webhook
      operationId: updateProfileWebhook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Webhook"
      responses:
        "200":
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

    delete:
      tags: [Profile]
      summary: Delete a webhook
      operationId: deleteProfileWebhook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      responses:
        "204":
          description: Webhook deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # -------------------------------- Strategy
  /strategy/alerts:
    get:
      tags: [Strategy]
      summary: List the account's trading alerts
      operationId: getStrategyAlerts
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: List of trading alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

    post:
      tags: [Strategy]
      summary: Create a new trading alert
      operationId: createStrategyAlert
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Alert"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /strategy/alerts/{alertId}:
    get:
      tags: [Strategy]
      summary: Get a specific trading alert
      operationId: getStrategyAlert
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: Alert details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

    put:
      tags: [Strategy]
      summary: Update a trading alert
      operationId: updateStrategyAlert
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AlertId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Alert"
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

    delete:
      tags: [Strategy]
      summary: Delete a trading alert
      operationId: deleteStrategyAlert
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "204":
          description: Alert deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Max items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      description: Items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    WebhookId:
      name: webhookId
      in: path
      required: true
      description: The unique ID of the webhook
      schema:
        type: integer
        format: uint64
        minimum: 1
    AlertId:
      name: alertId
      in: path
      required: true
      description: The unique ID of the trading alert
      schema:
        type: integer
        format: uint64
        minimum: 1
  responses:
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Sent too many request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
          example: Unauthorized
        details:
          type: object
          additionalProperties: true
    Password:
      type: string
      format: password
      writeOnly: true
      minLength: 8
      maxLength: 72
      description: |
        8-72 characters, including at least one uppercase,
        one lowercase, one digit, and one special character; no whitespace.
      pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z0-9])\S{8,72}$'
      example: Str0ngPassw0rd!
    Username:
      type: string
      minLength: 5
      maxLength: 20
      description: |
        5–20 characters; letters and numbers required;
        optional single dots or underscores allowed between characters;
        cannot start or end with dot/underscore; no consecutive dots or underscores.
      pattern: "^(?=.{5,20}$)(?!.*[_.]{2})[a-zA-Z0-9]+([._]?[a-zA-Z0-9]+)*$"
      example: member1
    Fullname:
      type: string
      minLength: 5
      maxLength: 200
      example: Nguyễn Vũ Hoàng Anh
    Email:
      type: string
      pattern: '^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}$'
      example: user1@example.com
    PhoneNumber:
      type: string
      description: the phone number of a created account
      pattern: '^\d{6,15}$'
      example: "0123456789"
    Role:
      type: string
      enum:
        - none
        - admin
        - member
      example: member
      default: member
    Account:
      type: object
      required: [username, email, role, fullname]
      properties:
        username:
          $ref: "#/components/schemas/Username"
        fullname:
          $ref: "#/components/schemas/Fullname"
        email:
          $ref: "#/components/schemas/Email"
        phoneNumber:
          $ref: "#/components/schemas/PhoneNumber"
        role:
          $ref: "#/components/schemas/Role"
    SignupRequest:
      type: object
      required: [account, password]
      properties:
        account:
          $ref: "#/components/schemas/Account"
        password:
          $ref: "#/components/schemas/Password"
    AccessTokenResponse:
      type: object
      additionalProperties: false
      required: [token, exp]
      properties:
        token:
          type: string
          description: JWT access token (store in memory; avoid localStorage if possible)
          example: eyJhbGciOi...
        exp:
          type: integer
          format: int64
          minimum: 1
          description: >
            Unix timestamp (seconds since epoch) when the access token expires.
          example: 1706812800
    SignupResponse:
      type: object
      additionalProperties: false
      required: [account, accessToken]
      properties:
        account:
          $ref: "#/components/schemas/Account"
        accessToken:
          $ref: "#/components/schemas/AccessTokenResponse"
    SigninRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          $ref: "#/components/schemas/Username"
        password:
          $ref: "#/components/schemas/Password"
        isRememberMe:
          type: boolean
          default: false
          description: If true, server may issue longer refresh token lifetime.
    SigninResponse:
      type: object
      additionalProperties: false
      required: [account, accessToken]
      properties:
        account:
          $ref: "#/components/schemas/Account"
        accessToken:
          $ref: "#/components/schemas/AccessTokenResponse"
    RefreshResponse:
      type: object
      additionalProperties: false
      required: [accessToken]
      properties:
        accessToken:
          $ref: "#/components/schemas/AccessTokenResponse"

    # -------------------------------- Profile
    ProfileMeResponse:
      type: object
      additionalProperties: false
      required: [account]
      properties:
        account:
          $ref: "#/components/schemas/Account"
    UpdateProfileMeRequest:
      type: object
      additionalProperties: false
      properties:
        fullname:
          $ref: "#/components/schemas/Fullname"
        email:
          $ref: "#/components/schemas/Email"
        phoneNumber:
          $ref: "#/components/schemas/PhoneNumber"
    UpdatePasswordRequest:
      type: object
      additionalProperties: false
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          $ref: "#/components/schemas/Password"
        newPassword:
          $ref: "#/components/schemas/Password"
    Webhook:
      type: object
      additionalProperties: false
      required: [name, url]
      properties:
        id:
          type: integer
          format: uint64
          minLength: 1
          maxLength: 26
          pattern: "^(?=.{5,20}$)(?!.*[_.]{2})[a-zA-Z0-9]+([._]?[a-zA-Z0-9]+)*$"
          readOnly: true
          description: Server-generated unique ID
          example: 42
        name:
          type: string
        url:
          type: string

    # -------------------------------- Strategy
    Timeframe:
      type: string
      enum:
        - D1
        - W1
        - M1
      example: D1
    Indicator:
      type: string
      enum:
        - close
        - bollinger_bands
        - rsi
        - ma10
        - ma50
        - ma100
        - ma200
      example: rsi
    Operator:
      type: string
      enum:
        - greater_than
        - less_than
        - crossing_up
        - crossing_down
        - crossing
      example: greater_than
    Trigger:
      type: string
      enum:
        - once
        - every
      example: once
    Alert:
      type: object
      additionalProperties: false
      required: [timeframe, symbol, indicator, operator, trigger, exp]
      properties:
        id:
          type: integer
          format: uint64
          readOnly: true
          description: Server-generated unique ID
          example: 1
        timeframe:
          $ref: "#/components/schemas/Timeframe"
        symbol:
          type: string
          minLength: 1
          maxLength: 20
          description: The stock ticker symbol
          example: AAPL
        indicator:
          $ref: "#/components/schemas/Indicator"
        operator:
          $ref: "#/components/schemas/Operator"
        trigger:
          $ref: "#/components/schemas/Trigger"
        exp:
          type: integer
          format: int64
          minimum: 1
          description: >
            Unix timestamp (seconds since epoch) when the alert expires.
          example: 1706812800
        message:
          type: string
          maxLength: 500
          description: Optional message for the alert
          example: RSI crossed above 70
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: When the alert was created
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          description: When the alert was last updated